Add experimental (forced) support to Wayland
Commit flags forces Wayland. Chromium has --ozone-platform-hint flag, which can be set to "auto" to automatically prefer Wayland over X11. Unfortunately, such flag is not available in the older Chromium that is packaged with VSCode. To this end, it is not safe (yet) to push this commit over Flathub official repos!